requirements.txt
python-telegram-bot==20.7
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
json5==0.9.14

